# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: API
  version: 1.0.0
paths:
...
  /users:
    get:
      summary: Получить всех пользователей
      tags:
        - users
      operationId: GetUsers
      responses:
        '200':
          description: Список пользователей
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UsersResponse'
    post:
      summary: Добавить нового пользователя
      operationId: CreateUser
      tags:
        - users
      requestBody:
        description: Новый пользователь
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUser'
      responses:
        '200':
          description: Новый пользователь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersResponse'
        '400':
          description: Некорректный запрос
        '422':
          description: Пользователь с таким email уже существует
        '500':
          description: Ошибка добавления пользователя
  /users/{id}:
    get:
      summary: Получить данные пользователя
      operationId: GetUserById
      tags:
        - users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: uint
      responses:
        '200':
          description: Данные пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersResponse'
        '404':
          description: Пользователь не найден
        '500':
          description: Ошибка при получении пользователя
    patch:
      summary: Обновить данные пользователя
      operationId: UpdateUser
      tags:
        - users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: uint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUser'
      responses:
        '200':
          description: Обновленный пользователь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersResponse'
        '400':
          description: Некорректный запрос
        '422':
          description: Пользователь с таким email уже существует
        '500':
          description: Ошибка обновления данных пользователя
    delete:
      summary: Удалить пользователя
      operationId: DeleteUserById
      tags:
        - users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: uint
      responses:
        '204':
          description: Пользователь успешно удален
        '404':
          description: Пользователь не найден
        '500':
          description: Ошибка удаления пользователя
components:
  schemas:
...
    UsersResponse:
      type: object
      properties:
        id:
          type: integer
          format: uint
        email:
          type: string
          format: email
      required:
        - id
        - email
    CreateUser:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    UpdateUser:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string